Moved functions into their own php file. Moved FontAwsome to a sub directory.
